Combined all the commits
Now, the login screen made by Lyn, displays correctly on the app. I seperated the login screen view to a distinct activity instead of a fragment, as it runs independently from MainActivity. Additonally, the app does NOT export MainActivity by default but only the LoginActivity; so it won't expose any valuable personal information to outside apps.